MB-61666 Allow unqualified references to target in RETURNING clause
This bug is a "regression" from the fix for MB-61485. When we formalize
the RETURNING clause of a MERGE statement, we now use a formalizer that
allows references to both target and source, such that references to
the source are properly handled. However a side effect of that change is
that expressions in the RETURNING clause must now be fully qualified,
since we now have 2 allowed aliases (previously it can be unqualified
since we only allow target alias).

To maintain backwards compatibility, i.e. still allowing unqualified
references to the target keyspace, we modified the formalizer used,
such that it still has 2 allowed aliases (source + target), but it
also sets the "expected" keyspace to the target keyspace, such that
any unqualified reference is formalized to the target keyspace (same
behavior as before), but any reference to the source keyspace must
be fully qualified.

Change-Id: Ia74f7a5c10e913affced69b66d1c13233684bc9a
Reviewed-on: https://review.couchbase.org/c/query/+/209158
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>