MB-61721 Allow nested array index key in simple form to be used
   ... for nested UNNEST operations

An leading array index key can be used for UNNEST scan, and when
the binding variable name used in the array index key definition
is different than the UNNEST alias, we rename the array index key
definition to match the UNNEST alias, such that we can proceed
with sargability determination of the array index key.

However, currently this rename of binding variables does not work
properly for the nested array index key, if the nested array
index key is defined in the simple form, i.e., "ALL <variable>"
instead of "ALL ARRAY v FOR v IN <variable> END". Because of this,
during index selection for nested UNNEST operations, we disable
an index for consideration if the nested array index key is in
the simple form.

The fix include:

 - modify the renaming process such that for simple form of the
   array index key, we also perform the rename operation on the
   arraymapping (which was missed before)
 - during index selection for nested array index keys, allow
   a nested array index key if it is in the simple form

Change-Id: I2b3ce932546e2a46beb2a85355dc2458ef376688
Reviewed-on: https://review.couchbase.org/c/query/+/209330
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker