MB-61707 Allow unnest scan for UNNEST operation without explicit alias
For UNNEST operation we consider unnest scan using an index with leading
array index key. This however currently does not work properly if the
UNNEST does not have an explicit alias (i.e. missing AS <alias>). This
is true for both the simplified and the expanded form of array index key.
This behavior has not changed since the beginning.

To fix this, instead of using the explicit alias (AS clause), we use
implicit alias as well, in case of missing explicit alias, during
unnest scan considerations.

Change-Id: I45c63f4a0b807a4deeda2da64233f611dec9a84c
Reviewed-on: https://review.couchbase.org/c/query/+/209344
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker