MB-61777 Guard against infinite recursion
This defect is a regression from the fix for MB-61620, whose
fix introduced a mechanism for calculating selectivity for
a few expression types (Not, Between, Like) which can be
transformed potentially. The issue is that if the expression
cannot be transformed, then we may get into infinite
recursion. Previously an attempt was made to limit infinite
recursion by checking whether the transformed expression is
equivalent to the original expression, however the function
EquivalentTo() returns false for a Subquery expression,
which then still results in infinite recursion.

To fix this, we now:

 - check for expression pointer matching in addition to
   EquivalentTo() for checking for equivalent expressions
 - add a recursion level check as well, such that we bail
   out if recursion level exceeds a predefined limit
 - the recursion check only applies to the case where we
   attempt to do expression transformation; in other
   situations where we also recurse, e.g. in case of AND,
   OR, or IN expressions, we also call the same function,
   however in those cases we don't use (potentially) the
   same expression in recursion and thus no issue of
   infinite recursion.

Change-Id: I26bce1618cb9edcb4124d7c04d456b40926e1a00
Reviewed-on: https://review.couchbase.org/c/query/+/209530
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>