MB-61842 Allow early order even if limit cannot be pushed to early order
This defect is a regression from the fix for MB-57666, where an
extra change was made to undo early order in presence of extra
filters, in which the limit cannot be pushed to the early order.
The rationale at the time was that "there is no clear advantage
for early order if limit cannot be pushed to the order". As it
turns out, there is an advantage to perform the order early,
even if the limit cannot be pushed to the order. The limit at
the end of the query plan can stop the entire plan once the
limit has been reached, and if the order is done early, since
things are pipelined after the order, we could end up fetching
a lot less number of documents in the Fetch operator. If the
order is done at the end of the plan, since Order is a
stop-n-go operator, we will need to fetch all qualifing index
keys from the index scan and feed the documents to the Fetch.
If the Order happens early (right after the index scan) then
as indicated earlier we could stop the plan early and thus
not Fetch as many documents.

The fix is basically undoing part of the change from the fix
for MB-57666. This should not affect the rest of the changes
for MB-57666 where we only force plans with order in case of
EXACTSPANS from the index scan.

Change-Id: Idbdfda2df665a52e64f21b910f1b18dcd2e92146
Reviewed-on: https://review.couchbase.org/c/query/+/209792
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/209793