MB-60646 Augment "No index available" message...
... to note checking the index service is available as a cluster may be
created without the service and the message suggests an index creation
statement.  (Done this way so that the primary message text is
unchanged.)

Detect when there are no Index nodes and don't attempt to create a
primary index on the system collection in this case.

Also forcibly enable sequential scans for the system collection. In the
absence of a GSI indexer we do still need to be able to scan our
metadata.

Change-Id: If5bd06302015bc9b53fa9d36bedf2dfcbdaff04b
Reviewed-on: https://review.couchbase.org/c/query/+/209825
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker