MB-61870 Ensure array id is generated for internally generated prediates
As part of the fix for MB-61549, we now require an ANY/ANY AND EVERY
predicate to have a valid array id in order to use it to sarg an
index key. The array id is generated early in the planning process.
The issue with this bug is that for internally generated predicates,
e.g. for dynamic index, we currently don't have this array id
generation and thus it errors out.

The fix is to ensure that for situations where we use internally
generated predicates, we also assign array id first. The assignment
of array id is only necessary if it has not been previouslly
assigned (e.g. if we use the same/copy of same predicate multiple
times in different planner paths).

Change-Id: I90ad0665c70c79c00691910d0216fb142760611d
Reviewed-on: https://review.couchbase.org/c/query/+/209963
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/209966