MB-62059 Proper generation of array index key in index advisor
The index advisor generates array index keys by evaluating the
ANY clause used in the query. For this bug, a non-traditiional
ANY clause is used:

  ANY i IN [1,2] SATISFIES i = id END

note that teh binding is referencing a constant array construct.
The index advisor is not handling this form well and generates
an invalid syntax for the array index key.

Several changes are made to address this issue:

 - add a "phase" when gathering array index construct, such
   that we can differentiate between the "binding" and the
   "satisfies" clauses of the ANY expression
 - after gathering from the bindings, if there is no
   "indexName" generated, i.e. the binding is not referring
   to the source keyspace, then bail out
 - when handling constants, only return the constant
   expression if the phase is in "satisfies", i.e. ignore
   the constants if handling "bindings"

After these changes, the index advisor now generates a
message indicating no index recommendation is available.

Change-Id: I82413d65dce32a0230a95f5668d05831b3edd6d1
Reviewed-on: https://review.couchbase.org/c/query/+/210612
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker