MB-62073 Forward port fix
Although 7.6 behaves differently, the fix is still valid in
7.6 as well.

In 7.6, we did not eliminate the 3rd index (without proper
cost) since the index has an index filter (which was not
present in 7.2, due to fix of MB-56616 in 7.6 only). Thus
we generate an intersect scan with 3 indexes in 7.6, but
since one of the indexes does not have proper cost, we
do perform early termination at runtime instead of waiting
for all index scans to finish, i.e. the slow performance
observed in 7.2 does not occur in 7.6. However the fix
in 7.2 is still valid.

The part about generating index spans is also different
in 7.6 vs 7.2. In 7.6 we generate index spans early on
and thus we could calculate cost even when shortest is
false.

Change-Id: If4293f78cf6cf93626314487b681dbe491278695
Reviewed-on: https://review.couchbase.org/c/query/+/210653
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/210661