MB-62073 Allow CBO to choose intersect scan ...          ... if an index without proper cost has been eliminated
When CBO is in effect, during index selection we calculate a cost
for each sargable index in order to choose the best index path.
If one (or more) of the sargable index(es) cannot be properly
costed (e.g. lack of appropriate statistics, or a bug/deficiency
in the code), then we turn CBO off. In this bug what happened
was that there are 3 sargable indexes, 2 of which have proper
costs but 1 does not. We turned CBO off (in the function) and
used rules to eliminate the index without proper cost. At the
end we generate an intersect scan with the two remaining indexes.
The problem is that the two indexes both have proper cost, and
at runtime we have a rule that says if the intersect scan is
chosen by cost (indicated by available cost for the intersect
scan), then we don't do early termination and let both index
scans finish. Since the two indexes both have proper cost, we
incorrectly determined at runtime that the intersect scan is
chosen by CBO while in reality it is not. Since one of the
indexes scans >95% of the index we timed out at runtime.

To fix this, we now "reset" useCBO in the function such that
when intersect scan is considered, we check and see whether
all remaining indexes have proper cost, and if so, still let
CBO choose the intersect scan, even if we've turned CBO off
in the function previously.

Also avoid calculating cost in case of "shortest == false",
since in that case we've not generated the index spans
and thus there is no point in trying to calculating cost
(this part behaves differently in 7.2 vs 7.6).

Change-Id: If4293f78cf6cf93626314487b681dbe491278695
Reviewed-on: https://review.couchbase.org/c/query/+/210653
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>