MB-62157 Ensure a covered ANY predicate remains covered ...          ... after covering transformation
In the problematic query we have a SubqueryTerm on the right-hand
side of a join, and there is an ANY predicate inside the subquery.
The subquery can be covered with an array index. The issue is
that the SubqueryTerm is planned multiple times during the planning
process. This is more obvious with join enumeration, but could
happen even if CBO is turned off.

The issue is that the first time the SubqueryTerm is planned, we
determine that covering is possible, and then proceed with cover
transformation for the subquery AST tree. Subsequent planning of
the SubqueryTerm now sees the transformed "cover" expression, and
for most expressions this works fine, i.e. does not interfere with
covering determination (since we just get the embedded expression
from the "cover" expression, see CoveredBy() function for Cover
expression, in expression/index_cover.go). However for ANY
predicate we use a special process (through filter covers) for
covering purpose, and this process currently does not handle
"cover" expressions, as a result all subsequent planning of the
SubqueryTerm incorrectly determines that covering cannot be done.

While this process is somewhat flawed (we should probably not
perform covering transformation the first time the SubqueryTerm
is planned), however to change this involves fairly complicated
infrastructual change and is probably not desiable for 7.6 at
this point. Instead a simple change to allow "cover" expression
to work for ANY predicate when determining filter covers is
done to allow proper determination of covering (with ANY
predicate) if the expression has already gone through cover
transformation.

Change-Id: Ib0c01a1250071150636fb0aa3a5f8dca91d48d93
Reviewed-on: https://review.couchbase.org/c/query/+/210933
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>