MB-62216 Handle OFFSET properly in early order
The issue in this defect is with the handling of OFFSET in
early order when the order also has a partialSortTermCount,
i.e. the index provides part of the total order.

The current code assumes that when partialSortTermCount is
present, then the OFFSET is handled by the SORT operator.
This is true for the "regular" SORT operator (typically at
the end of the plan), but not always true for early order,
since in case the index scan does not have EXACTSPANS we
cannot push the OFFSET to the early SORT operator, in
which case the OFFSET cannot be handled by early order.

To fix this we need to remove the assumption for
partialSortTermCount in early order, and just let existing
logic takes care of determining whether OFFSET is already
handled (which is based on this.offset == nil).

Change-Id: I843849aa5bb91276fe4a9f1c4ab0d14587a5d586
Reviewed-on: https://review.couchbase.org/c/query/+/211084
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker