MB-62274 Proper handling of outer join during join enumeration
The problem with this defect is a combination of the following:

 - an outer join is performed, with the join condition on meta().id
   of the subservient side:
 - the join condition on meta().id involved a CASE statement
 - there is an index with meta().id as the leading index key on the
   subservient side
 - join enumeration is in play

The main issue here is that we incorrectly derived a predicate from
the ON-clause filter during join enumeration, when generating
single-keyspace plans, in the form of: meta(r).id IS NOT NULL. This
should not have happened since the join is an outer join.

The fix is to avoid deriving predicate from ON-clause of an outer join
during join enumeration (this is not an issue if join enumeration is
not in play since we already have a check for !join).
 -

Change-Id: I7c8ea0769c00b93053f0090880b843ab3e96b6fa
Reviewed-on: https://review.couchbase.org/c/query/+/211168
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>