MB-62366 Always track RAW projections...
... of AnnotatedValue objects.  This is so that if used further
downstream in the statement (or say cached) it isn't release
prematurely.

This may mean for some cases AnnotatedValue objects are not recycled but
this is acceptable in place of more complex tracking.  (The GC load
shouldn't be so great as to be an issue as a result of this.)

Change-Id: If7274396f70e8abb9149e27619f30f6406f45eee
Reviewed-on: https://review.couchbase.org/c/query/+/211422
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker