MB-62373 Compact subquery hints array ...
... when removing optimizer hints for a SubqueryTerm

When a SubqueryTerm is used on the inner of a join, and
nested-loop join is used, then we put an ExpressionScan
on top of the subquery. As part of this process, we
remove optimizer hints from the SubqueryTermHints. The
current code just set the pointer to nil, which then
relies on nil pointer check for any subsequent access
to elements of the array.

Instead we now compact the array when an element is
being removed, such that we don't leave a nil pointer
in the array.

Change-Id: I2665ec8b08b0a7d57befb8c140128c714c0f6f44
Reviewed-on: https://review.couchbase.org/c/query/+/211484
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>