MB-61782: Update TLS configuration of Query's secure ports without...
closing and re-starting the secure listeners.

A callback is now set in the GetConfigForClient() method of the
listener's TLS configuration.

This callback returns the TLS config to be used for the TLS
handshake when when a ClientHello message is received.

Query will maintain a single TLS configuration for the encrypted port.
This configuration object will be returned by the callback mentioned
above.

Hence, any update to the port's TLS Configuration config returned by
the callback.

And the close and restart of the listener will no longer be needed
to load the new TLS configuration.

Change-Id: Ia75fc51c1ae8345f0b244fdbc9092c7811c0864f
Reviewed-on: https://review.couchbase.org/c/query/+/211656
Well-Formed: Restriction Checker
Tested-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>