MB-62394 Revise profile reporting.
The execution analyser now forces operators to report/accumulate their
times when it is run.  As phase switching is invoked outside the
operator itself, some simple serialisation is introduced in the phase
switching code.

The analyser call is moved to the start of profile reporting so that
reported times are as up-to-date as possible.  (Without we could report
say "High IO" with a negligible time in the "phaseTimes" and "ioTime"
fields, but a high servTime for e.g. the fetch operator.)

The redundant condition (p != server.ProfOff) was removed from profile
reporting too, just to clean-up the profile reporting.

As the analysis is run for all enabled profile settings, it is included
for all now too.

Further, corrects RequestLog.Format to include the waitTime, ioTime,
~analysis & sessionMemory fields which were omitted/lost in the merge of
435b322340037530c217429cd22f5615272b14a9.

Change-Id: I1a164d5330466b211e07bdcc1c6245158a0a74d3
Reviewed-on: https://review.couchbase.org/c/query/+/212104
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>