MB-60680 Change DOT_PRODUCT distance to be negated ...    ... DOT_PRODUCT similarity
In general vector distance is smaller for two vectors that are
close (or similar) and larger for two vectors that are not.
Of the vector distance metrics that we currently support, there
are two that are opposite to this general trend:
  - COSINE similarity, which is larger if two vectors are similar
  - DOT_PRODUCT similarty, which is larger if two vectors are
    similar

Of the two, we already handle COSINE similarity, by doing the
calculation as (1 - COSINE_SIMILARITY) to be COSINE_DISTANCE.

For DOT_PRODUCT distance, we also need to modify the distance
calculation to be negated DOT_PRODUCT similarity, such that
smaller number means more closer (or similar) vectors.

Note that both of these manipulations are the same as what
Oracle does. (pgvector also does it this way).

The fix include:

 - change calculation of vector distance calculation to use
   negated number for DOT_PRODUCT
 - undo changes from yesterday that use DESC for ordering for
   DOT PRODUCT
 - change (APPROX_)COSINE_SIM_DIST to just (APPROX_)COSINE_DIST
   to be more precise (and also inline with what Oracle uses)

Change-Id: I75afca3f74c20cf60152358fc748deca457c17f8
Reviewed-on: https://review.couchbase.org/c/query/+/212173
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>