MB-60680 Fix build issue
The trinity/cypher branch is missing the change:

  https://review.couchbase.org/c/query/+/197164

which includes a lot of files, but the only thing needed
to fix the build is a new FunctionBase.Init() function.
Thus just add that one function for now, and maybe later
we can backport the rest of that change.

Also changed the version in index.go from 8.0.0 to 7.7.0.

The build now works:
  - query and query-ee on cypher branch
  - n1fty and eventing-ee on trinity branch
  - rest are from morpheus build

Change-Id: Id1b55ece68bd757c08d64dd47459ada2ae0938b1
Reviewed-on: https://review.couchbase.org/c/query/+/212241
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>