MB-62631 Remove sequential scan privilege setting in planner
The new RBAC privilege for sequential scan was introduced in
trinity and went through several revisions. MB-60129 added
the capability for the planner to check for the privilege and
avoid generating sequential scan plans; part of that change
was reverted in MB-60291 due to issues with upgrade scenarios.
There are a few changes left over that still adds extra
privileges in the planner for sequential scans. These settings
could be problematic in cases where multiple plans could be
generated for the same KeyspaceTerm (including a sequential
scan along the way) but we end up choosing a non-sequential
scan plan for the keyspace.

For this bug, the ADVISE statement goes through the planner
and the planner generated a sequential scan plan, which adds
the extra privilege to be checked at runtime. Subsequent
runtime check resulted in a 13014 error for sequential scan
although the Advise statement itself does not involve any
sequential scan.

The fix is to remove the remaining setting of extra privileges
in the planner, and rely on the new visitor function to gather
all sequential scan privileges by walking the final plan. This
will then allow the Advise statement to produce meaningful
result for a user that does not have sequential scan privilege.

Note that a somewhat similar issue exists today for EXPLAIN
statement as well, i.e. EXPLAIN of a statement also generates
a 13014 error currently. With the fix the EXPLAIN statement
also produces an actual plan which uses sequential scan. If
the user attempts to execute the statement, 13014 error will
result then.

Potential enhancement to EXPLAIN in this case can be
considered in the future, e.g. if user does not have the
sequential scan privilege, we can potentially generate an
warning for the EXPLAIN statement.

Change-Id: Ie08e4faa874368bc56f96a3f013189b835b0d337
Reviewed-on: https://review.couchbase.org/c/query/+/212415
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>