MB-60680 Support non-squared version of EUCLIDEAN/L2 ...    ... as distance metric for vector functions
The indexer uses FAISS library for vector index processing and
FAISS only support the SQUARED versions of L2/EUCLIDEAN.
Previously we decided that the approximate vector function
APPROX_VECTOR_DISTANCE only supports the SQUARED versions of
L2/EUCLIDEAN. However, this may impose unnecessary hardship on
usage of these functions since users are familiar with L2 and
EUCLIDEAN. In addition, our mobile offering also supports L2
and EUCLIDEAN as vector distance metric.

This is the first part of adding support for the non-SQUARED
versions of EUCLIDEAN/L2 as distance metric for
APPROX_VECTOR_DISTANCE (note they are already supported for
VECTOR_DISTANCE). What we do is to internally translate the
non-SQUARED version to the SQUARED version before passing to
indexer, such that the indexer still deals with the SQUARED
versions, and we deal with the conversion to the non-SQUARED
version.

The fix includes:

 - add an expression flag for ORDER BY terms
 - translate non-SQUARED version to SQUARED version of L2 and
   EUCLIDEAN before passing to indexer
 - allow non-SQUARED version of L2/EUCLIDEAN in
   APPROX_VECTOR_DISTANCE
 - add function to replace the metric in APPROX_VECTOR_DISTANCE
   function (a new function with the non-SQUARED version to the
   SQUARED version)
 - add CoveredBy() function for APPROX_VECTOR_DISTANCE function
   to allow covering of the non-SQUARED version by the SQUARED
   version, if the function is in the ORDER BY clause
 - when determining index covering, replace the non-SQUARED
   version with SQUARED version (which is what indexer provides)

With these changes, if APPROX_VECTOR_DISTANCE only appears in
the ORDER BY clause it'll behaves the same as before, if it is
also appearing in other parts of the query (e.g. projection)
then we'll need a FETCH and recalculate the distance.

We'll likely need additional work on index projection and
order pushdown to make this work fully.

Change-Id: I4c24c8e198832b542ba3e9ed7d30ee3f93850480
Reviewed-on: https://review.couchbase.org/c/query/+/212484
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>