MB-13085 Remove GROUP BY spilling and react to ...
...memory tracking failures in spilling (ORDER BY) directly.

GROUP BY spilling in ineffective except for the case where we have
sufficient groups to exceed pipeline-batch sizing, and no "dam"
downstream in the pipe (e.g. an ORDER BY).  In the absence of these
conditions we end up having to have all complete groupings in memory at
the same time so all spilling achieves is delaying the inevitable
out-of-quota error.  Since the spilling offers no real benefits here and
adds complexity to the code, it has been decided to remove it.

To react better to quota errors when using spilling (ORDER BY, etc.),
the memory tracking function provided can now return an error which will
halt further processing (in addition to typically failing the request
with the error).

Further, this function is propagated throughout the reading of spilled
values enabling us to stop materialising them sooner, reducing the
over-shoot that indicates the quota has been exhausted.  (Better keeps
actual memory use in check.)

Change-Id: I707de3901e7be365ed450133f30c61c6c95c5f52
Reviewed-on: https://review.couchbase.org/c/query/+/212792
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>