MB-60680 Proper resetting of order pushdown
When Ann is present in ORDER BY clause, we initially attempt to
pushdown the order to index if possible. Later on, when we
determine that LIMIT cannot be pushdown for whatever reason, we
need to remove the vector predicate from the ORDER pushdown
since in this case the indexer cannot provide ordering for the
vector index key. Previous code was attempting to convering
ORDER pushdown to partial ORDER pushdown, however that is not
always correct since there are other factors affecting the
(partial) ORDER pushdown such as equality predicates on an
index key. This may result in incorrect calculation for number
of partial order keys.

To fix this, we now go through the same logic for determining
index order again, with an additional setting that indicates
the vector index key should not be considered for ordering.

Change-Id: If10ae0f499f35ccbd2b085e37429efeb7e76da46
Reviewed-on: https://review.couchbase.org/c/query/+/212809
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>