MB-60680 Allow EUCLIDEAN and L2 to be compatible
The vector index metric EUCLIDEAN and L2 are the same in
index distance calculation. This also include the SQUARED
version of the metric EUCLIDEAN_SQUARED and L2_SQUARED.
Thus an index defined with one can be used for a search
with another.

Make EUCLIDEAN and L2 metrics compatible with each other
during index selection.

Change-Id: I9ccac69c4257c95d234ed4d5c70d96874e43afba
Reviewed-on: https://review.couchbase.org/c/query/+/212825
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>