MB-60680 Lift restriction on ORDER pushdown for vector index key
Currently if a vector index key is not a leading index key, we only
allow ORDER pushdown for vector index key if all leading index keys
have equality predicate only.

In recent discussions with index team we decided that we can push
down ORDER for vector index key as long as LIMIT can also be
pushed down. E.g. if a leading index key has a range predicate or
an IN-clause, we can still pushdown ORDER for vector index key
as long as LIMIT can also be pushed.

This checkin removes the restriction on equality predicate only
for leading index keys.

Change-Id: I9ee402b480b643cd385261ade1ebdf97d324332d
Reviewed-on: https://review.couchbase.org/c/query/+/212912
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>