MB-62824 Revert GC hint changes
In MB 61526 an attempt was made to hint to the GC WITH binding array
values were no longer needed to encourage freeing of memory sooner.

This was flawed as it involves altering the value however this may be a
shared copy since the expression package caches values for constants and
returns the same value instance for multiple evaluations.  As the result
from expression evaluation may be shared it must always be treated as
being shared and therefore mustn't be altered.

Change-Id: Ia63c59be37047c114dc61a924ee238a02ede3061
Reviewed-on: https://review.couchbase.org/c/query/+/213070
Well-Formed: Restriction Checker
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>