MB-60680 Don't unset SKIP ORDER flag for vector index key
If LIMIT is not pushed down to indexer, then we cannot push
ORDER (for the vector index key) to indexer either. Previously
we added a new flag to index entry to avoid pushing ORDER to
indexer for vector index key. This flag was unset after we
obtained partialOrderCount though. What happened was that later
on when we try to get index order positions we incorrectly
included the vector index key.

The fix is to avoid resetting the SKIP ORDER flag.

Change-Id: Ia7bd7b2d428bf58e2c76a144ac2dac03070b5857
Reviewed-on: https://review.couchbase.org/c/query/+/213255
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>