MB-60680 Adjust vector index key projection
If a vector index key is projected in index projection, the index
will need to calculate vector distance and return to query, this
is an cpu-expensive operation. Current code will include the vector
index key in index projection if there is a suitable vector index.
This can potentially require unnecessary vector distance calculation
for certain queries.

Modify the criteria for vector index key projection such that a
vector index key is projected if:
  - ORDER is not pushed down to indexer
  - equivalent ANN expression is present in query other than the
    ORDER BY clause

With this fix, well omit the vector index key in index projection
if:
  - ORDER is pushed to the indexer (already done)
  - there are no other references to the ANN expression aside from
    the ORDER BY clause

Change-Id: I6150147575d699c8d22191e7bfc17b4628b94009
Reviewed-on: https://review.couchbase.org/c/query/+/213316
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>