MB-62996 Add semantic check for query vector ...     ... if query is not using vector index
If a vector index is used, then we have logic at execution time
to check the validity of the query vector. However, if the query
has vector distance function but is not using a vector index,
then currently there is no checking for a valid query vector. In
case the query vector is not valid, we simply returns a NULL
value as the result of expression evaluation. This is somewhat
inconsistent since depending on whether a vector index is
available, the same query may return an error in one case but
NULL value in another.

We add a check for valid query vector as part of the expression
evaluation. This is done early on if the query vector is a
constant, otherwise it is done at expression evaluation time
(execution time). This execution time check should only trigger
if a vector index is not used, otherwise it should have been
caught earlier.

Change-Id: I09e6a9f7238cb4889bfef40ec01f97fe2bd34070
Reviewed-on: https://review.couchbase.org/c/query/+/213670
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>