MB-63016 skip CASE expression in DNF transformation
The issue with the customer scenario is an index definition where
an index key has a CASE statement, and in one arm of the CASE
WHEN condition there is a not equal (<>) expression. When the
planner gets index keys it does DNF transformation on the index
key expressions, since we do the same DNF transformation for the
WHERE/ON clause expression in order to do index selection.
However in the customer's case the same CASE expression is used
in the projection clause, and since we do not do DNF transformation
on projection clause expressions, we do not properly match the
CASE expression and thus determined that covering is not feasible.

Since the main purpose for performing DNF transformation is index
selection, and we don't really do index selection with subclauses
of a CASE expression, there is really no need to do the DNF
transformation for a CASE expression.

We now skip DNF transformation for a CASE expression. This should
(hopefully) not affect index selection since the index key
expressions are kept in its original form, and we only do the DNF
transformation after we get the index key expressions in the
planner (i.e. as part of query processing).

Change-Id: Ief8259e95bc1f3765860174452757b9307ba747d
Reviewed-on: https://review.couchbase.org/c/query/+/213745
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker