MB-63060 Avoid unnecessary function body creation during migration
For UDF migration, currently it uses the same code logic and normal
UDF creation, which involves creation of function name, function
body then save the necessary info in system collection. The issue
with migration is that if an UDF has a nested call to another UDF,
then the current code logic will try to create a function body for
the UDF being migrated, which then attempts to load info for the
nested UDF, which will wait for migration to complete, i.e. it'll
deadlock on itself.

To avoid this, we now omit the step for function body creation,
since it is not really required for migration purpose (all we do
is to turn around and create a string representation for the
function body, which is already available from the original
entry). Instead we just obtain the function entry from the
original metakv location, and directly save the function entry
into the new location (system collection). This avoids the
unnecessary middle step of creating then linearing the function
body.

Change-Id: Iab72cf816b3f646d269f5e7d42e2537259747d32
Reviewed-on: https://review.couchbase.org/c/query/+/213894
Well-Formed: Restriction Checker
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
(cherry picked from commit 86ad4f3659d4872076512b7182cc30a89d84470b)
Reviewed-on: https://review.couchbase.org/c/query/+/213788
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>